fix CodeQL: cast multiplication operands to prevent widening overflow in 8 files

+++ src/generator_common_aarch64.c
@@ -4478,7 +4478,7 @@ void libxsmm_generator_load_32x32_aarch64_sme( libxsmm_generated_code* io_genera
                                                     (l_en==0) ? i_gp_reg_addr : i_gp_reg_help,
                                                     LIBXSMM_AARCH64_GP_REG_X11,
                                                     (l_en==0) ? i_gp_reg_addr : i_gp_reg_help,
-                                                    i_ldc * 4 );
+                                                    (unsigned long long)i_ldc * 4 );
     }
 
     /* mov to za register */
@@ -4547,7 +4547,7 @@ void libxsmm_generator_load_64x16_aarch64_sme( libxsmm_generated_code* io_genera
                                                       i_gp_reg_addr,
                                                       LIBXSMM_AARCH64_GP_REG_X11,
                                                       i_gp_reg_addr,
-                                                      i_ldc * 4 );
+                                                      (unsigned long long)i_ldc * 4 );
       }
       for( l_en = 16; l_en < ((i_n_blocking > 7) ? 20 : (i_n_blocking + 12)); l_en++){
         libxsmm_aarch64_instruction_sme_mov( io_generated_code,
@@ -4561,7 +4561,7 @@ void libxsmm_generator_load_64x16_aarch64_sme( libxsmm_generated_code* io_genera
                                                       i_gp_reg_addr,
                                                       LIBXSMM_AARCH64_GP_REG_X11,
                                                       i_gp_reg_addr,
-                                                      i_ldc * 4 );
+                                                      (unsigned long long)i_ldc * 4 );
       }
     } else {
       /* load second block of C */
@@ -4577,7 +4577,7 @@ void libxsmm_generator_load_64x16_aarch64_sme( libxsmm_generated_code* io_genera
                                                       i_gp_reg_addr,
                                                       LIBXSMM_AARCH64_GP_REG_X11,
                                                       i_gp_reg_addr,
-                                                      i_ldc * 4 );
+                                                      (unsigned long long)i_ldc * 4 );
       }
       for( l_en = 16; l_en < ((i_n_blocking > 15) ? 20 : (i_n_blocking + 4)); l_en++){
         libxsmm_aarch64_instruction_sme_mov( io_generated_code,
@@ -4591,7 +4591,7 @@ void libxsmm_generator_load_64x16_aarch64_sme( libxsmm_generated_code* io_genera
                                                       i_gp_reg_addr,
                                                       LIBXSMM_AARCH64_GP_REG_X11,
                                                       i_gp_reg_addr,
-                                                      i_ldc * 4 );
+                                                      (unsigned long long)i_ldc * 4 );
       }
     }
     /* mov to za register */
@@ -4755,7 +4755,7 @@ void libxsmm_generator_store_32x32_aarch64_sme( libxsmm_generated_code* io_gener
                                                     (l_en==0) ? i_gp_reg_addr : i_gp_reg_help,
                                                     LIBXSMM_AARCH64_GP_REG_X11,
                                                     (l_en==0) ? i_gp_reg_addr : i_gp_reg_help,
-                                                    i_ldc * 4 );
+                                                    (unsigned long long)i_ldc * 4 );
     }
     l_n_count -= 16;
   }
@@ -4851,7 +4851,7 @@ void libxsmm_generated_store_64x16_aarch64_sme( libxsmm_generated_code* io_gener
                                                       i_gp_reg_addr,
                                                       LIBXSMM_AARCH64_GP_REG_X11,
                                                       i_gp_reg_addr,
-                                                      i_ldc * 4 );
+                                                      (unsigned long long)i_ldc * 4 );
       }
       for( l_en = 16; l_en < ((i_n_blocking > 7) ? 20 : (i_n_blocking + 12)); l_en++){
         libxsmm_aarch64_instruction_sme_mov( io_generated_code,
@@ -4865,7 +4865,7 @@ void libxsmm_generated_store_64x16_aarch64_sme( libxsmm_generated_code* io_gener
                                                       i_gp_reg_addr,
                                                       LIBXSMM_AARCH64_GP_REG_X11,
                                                       i_gp_reg_addr,
-                                                      i_ldc * 4 );
+                                                      (unsigned long long)i_ldc * 4 );
       }
     } else {
       /* load second block of C */
@@ -4881,7 +4881,7 @@ void libxsmm_generated_store_64x16_aarch64_sme( libxsmm_generated_code* io_gener
                                                       i_gp_reg_addr,
                                                       LIBXSMM_AARCH64_GP_REG_X11,
                                                       i_gp_reg_addr,
-                                                      i_ldc * 4 );
+                                                      (unsigned long long)i_ldc * 4 );
       }
       for( l_en = 16; l_en < ((i_n_blocking > 15) ? 20 : (i_n_blocking + 4)); l_en++){
         libxsmm_aarch64_instruction_sme_mov( io_generated_code,
@@ -4895,7 +4895,7 @@ void libxsmm_generated_store_64x16_aarch64_sme( libxsmm_generated_code* io_gener
                                                       i_gp_reg_addr,
                                                       LIBXSMM_AARCH64_GP_REG_X11,
                                                       i_gp_reg_addr,
-                                                      i_ldc * 4 );
+                                                      (unsigned long long)i_ldc * 4 );
       }
     }
   }
@@ -4954,7 +4954,7 @@ void libxsmm_generator_transpose_sme( libxsmm_generated_code * io_generated_code
                                                     i_gp_reg_addr,
                                                     LIBXSMM_AARCH64_GP_REG_X11,
                                                     i_gp_reg_addr,
-                                                    i_k * 4 );
+                                                    (unsigned long long)i_k * 4 );
     }
   } else {
     for( l_i = 0; l_i < i_n_blocking; l_i++){
@@ -4969,7 +4969,7 @@ void libxsmm_generator_transpose_sme( libxsmm_generated_code * io_generated_code
                                                       i_gp_reg_addr,
                                                       LIBXSMM_AARCH64_GP_REG_X11,
                                                       i_gp_reg_addr,
-                                                      i_k * 4 );
+                                                      (unsigned long long)i_k * 4 );
     }
   }
   /* mov to za register */
@@ -5161,7 +5161,7 @@ void libxsmm_generator_sme_transpose_64( libxsmm_generated_code * io_generated_c
                                                   i_gp_reg_addr,
                                                   i_help_register,
                                                   i_gp_reg_addr,
-                                                  i_k * 4 );
+                                                  (unsigned long long)i_k * 4 );
   }
 
   /* mov to tile */
@@ -5210,7 +5210,7 @@ void libxsmm_generator_sme_transpose_64( libxsmm_generated_code * io_generated_c
                                                   i_gp_reg_addr,
                                                   i_help_register,
                                                   i_gp_reg_addr,
-                                                  i_k * 4 );
+                                                  (unsigned long long)i_k * 4 );
   }
   if( l_blocking_phase[1] > 0 ){
     for( l_i = 0; l_i < 4; l_i++){
@@ -5387,7 +5387,7 @@ void libxsmm_generated_load_16x64_aarch64_sme( libxsmm_generated_code* io_genera
                                                    i_gp_reg_addr,
                                                    LIBXSMM_AARCH64_GP_REG_X11,
                                                    i_gp_reg_addr,
-                                                   i_ldc * 4 );
+                                                   (unsigned long long)i_ldc * 4 );
   }
 
   /* mov to za register */
@@ -5436,7 +5436,7 @@ void libxsmm_generated_load_16x64_aarch64_sme( libxsmm_generated_code* io_genera
                                                       i_gp_reg_addr,
                                                       LIBXSMM_AARCH64_GP_REG_X11,
                                                       i_gp_reg_addr,
-                                                      i_ldc * 4 );
+                                                      (unsigned long long)i_ldc * 4 );
     }
     /* mov to za register */
     for( l_i = 0; l_i < 4; l_i++){
@@ -5536,7 +5536,7 @@ void libxsmm_generated_store_16x64_aarch64_sme( libxsmm_generated_code* io_gener
                                                    i_gp_reg_addr,
                                                    LIBXSMM_AARCH64_GP_REG_X11,
                                                    i_gp_reg_addr,
-                                                   i_ldc * 4 );
+                                                   (unsigned long long)i_ldc * 4 );
   }
   /* mov tile 3-4 */
   if(i_n_blocking > 32){
@@ -5584,7 +5584,7 @@ void libxsmm_generated_store_16x64_aarch64_sme( libxsmm_generated_code* io_gener
                                                     i_gp_reg_addr,
                                                     LIBXSMM_AARCH64_GP_REG_X11,
                                                     i_gp_reg_addr,
-                                                    i_ldc * 4 );
+                                                    (unsigned long long)i_ldc * 4 );
     }
   }
 }
